IDs fixed

+++ app/src/main/java/com/example/progass2/Profile.java
@@ -1,23 +1,23 @@
 package com.example.progass2;
 
 public class Profile {
-    private long id;
+    private int id;
     private String name;
     private String surname;
     private float gpa;
 
-    public Profile(long id, String name, String surname, float gpa) {
+    public Profile(int id, String name, String surname, float gpa) {
         this.id = id;
         this.name = name;
         this.surname = surname;
         this.gpa = gpa;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
